+++ acmg_db/views/__init__.py
@@ -1,0 +1,8 @@
+from .generic_views import *
+from .import_views import *
+from .first_check_views import *
+from .second_check_views import *
+from .classification_views import *
+from .variant_views import *
+from .panel_views import *
+from .reporting_views import *
